OneSoko/OneSokoApp/requirements_enhanced.txt
# Enhanced OneSoko API Requirements
# Core Django and DRF
Django>=4.2.0,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.2.0,<6.0
django-cors-headers>=4.0.0,<5.0

# Performance and Caching
django-redis>=5.2.0,<6.0
redis>=4.5.0,<5.0
django-cacheops>=7.0.0,<8.0

# Database Optimization
django-debug-toolbar>=4.0.0,<5.0
django-extensions>=3.2.0,<4.0

# Security
django-ratelimit>=4.0.0,<5.0
django-axes>=6.0.0,<7.0
django-defender>=0.9.0,<1.0

# API and Filtering
django-filter>=23.0,<24.0
drf-yasg>=1.21.0,<2.0
django-rest-swagger>=2.2.0,<3.0

# Real-time and WebSockets
channels>=4.0.0,<5.0
channels-redis>=4.1.0,<5.0
daphne>=4.0.0,<5.0

# Monitoring and Logging
django-prometheus>=2.3.0,<3.0
sentry-sdk>=1.28.0,<2.0

# Image Processing
Pillow>=10.0.0,<11.0

# Utilities
python-decouple>=3.8,<4.0
celery>=5.3.0,<6.0
django-celery-beat>=2.5.0,<3.0
django-celery-results>=2.5.0,<3.0

# Development and Testing
pytest>=7.4.0,<8.0
pytest-django>=4.5.0,<5.0
factory-boy>=3.3.0,<4.0
coverage>=7.3.0,<8.0

# Production
gunicorn>=21.2.0,<22.0
whitenoise>=6.5.0,<7.0
psycopg2-binary>=2.9.0,<3.0

# Optional: Advanced Features
django-storages>=1.14.0,<2.0  # For S3 file storage
boto3>=1.28.0,<2.0  # AWS SDK
django-allauth>=0.54.0,<1.0  # Social authentication
django-rest-auth>=0.9.5,<1.0  # REST authentication 